Removed unused reader files
