+++ projects/openapi-reader/src/register.ts
@@ -1,8 +1,7 @@
 import { hash } from './hash'
-import { ReadContext, ReadInput } from './internalTypings'
+import { ReadContext } from './internalTypings'
 
-export function register(input: ReadInput<any>, context: ReadContext): void {
-  const { uri, data } = input
+export function register(data: any, uri: string, context: ReadContext): void {
   context.cache.uriToObject.set(uri, data)
   context.cache.objectToUri.set(data, uri)
   context.cache.objectToHash.set(data, hash(JSON.stringify(data)))
