--- projects/openapi-reader/src/utils/tick.ts
@@ -1,3 +1,0 @@
-export function tick() {
-  return new Promise((resolve) => setTimeout(resolve, 0))
-}
