--- projects/openapi-reader/src/validators/pathsObject.ts
@@ -1,3 +1,0 @@
-import { record, object, string } from '@oats-ts/validators'
-
-export const pathsObject = object(record(string(), object()))
