--- projects/openapi-reader/src/validators/recordOfObjects.ts
@@ -1,3 +1,0 @@
-import { object, record, string } from '@oats-ts/validators'
-
-export const recordOfObjects = object(record(string(), object()))
