--- projects/openapi-reader/src/validators/referenceObject.ts
@@ -1,3 +1,0 @@
-import { shape, object, string } from '@oats-ts/validators'
-
-export const referenceObject = object(shape({ $ref: string() }))
